api-ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
rembg==2.0.67
pillow==10.1.0
onnxruntime==1.23.2
